+++ crate/rt/src/commands.rs
@@ -1,10 +1,10 @@
 pub use self::{
-    diff_cmd::DiffCmd, ensure_cmd::EnsureCmd, state_current_cmd::StateCurrentCmd,
-    state_desired_discover_cmd::StateDesiredDiscoverCmd, state_discover_cmd::StateDiscoverCmd,
+    diff_cmd::DiffCmd, ensure_cmd::EnsureCmd, state_desired_discover_cmd::StateDesiredDiscoverCmd,
+    state_discover_cmd::StateDiscoverCmd, states_current_discover_cmd::StatesCurrentDiscoverCmd,
 };
 
 mod diff_cmd;
 mod ensure_cmd;
-mod state_current_cmd;
 mod state_desired_discover_cmd;
 mod state_discover_cmd;
+mod states_current_discover_cmd;