+++ crate/rt/src/commands/state_discover_cmd.rs
@@ -7,7 +7,7 @@ use peace_resources::{
 };
 use peace_rt_model::CmdContext;
 
-use crate::{StateCurrentCmd, StateDesiredDiscoverCmd};
+use crate::{StateDesiredDiscoverCmd, StatesCurrentDiscoverCmd};
 
 #[derive(Debug)]
 pub struct StateDiscoverCmd<E>(PhantomData<E>);
@@ -33,7 +33,8 @@ where
         cmd_context: CmdContext<'_, SetUp, E>,
     ) -> Result<CmdContext<WithStatesCurrentAndDesired, E>, E> {
         let (workspace, item_spec_graph, mut resources) = cmd_context.into_inner();
-        let states = StateCurrentCmd::exec_internal(item_spec_graph, &mut resources).await?;
+        let states =
+            StatesCurrentDiscoverCmd::exec_internal(item_spec_graph, &mut resources).await?;
         let states_desired =
             StateDesiredDiscoverCmd::exec_internal(item_spec_graph, &mut resources).await?;