+++ workspace_tests/src/rt.rs
@@ -1,5 +1,5 @@
 mod diff_cmd;
 mod ensure_cmd;
-mod state_current_cmd;
 mod state_desired_discover_cmd;
 mod state_discover_cmd;
+mod states_current_discover_cmd;