+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-* `StateCurrentCmd` to discover the current state.
+* `StatesCurrentDiscoverCmd` to discover the current state.
 * `StateDesiredDiscoverCmd` to discover the desired state.
 * `DiffCmd` to compute the difference between the current and desired states.
 * `EnsureCmd` to transform the current state into the desired state.